requirements.txt
streamlit
pandas
requests
matplotlib
geopy
folium
streamlit-folium
google-generativeai
scikit-learn